templateplus component drafted, firmware dir skeleton drafted, Kconfig restructured, arduino hooks ready
