+++ components/templateplus/templateplus.c
@@ -1,0 +1,66 @@
+#include "templateplus.h"
+
+static esp_err_t event_handler(void *ctx, system_event_t *event)
+{
+    switch(event->event_id) {
+    case SYSTEM_EVENT_STA_START:
+        esp_wifi_connect();
+        break;
+    case SYSTEM_EVENT_STA_GOT_IP:
+        xEventGroupSetBits(wifi_event_group, WIFI_CONNECTED_BIT);
+        break;
+    case SYSTEM_EVENT_STA_DISCONNECTED:
+        esp_wifi_connect();
+        xEventGroupClearBits(wifi_event_group, WIFI_CONNECTED_BIT);
+        break;
+    default:
+        break;
+    }
+    return ESP_OK;
+}
+
+static void wifi_init()
+{
+    tcpip_adapter_init();
+    wifi_event_group = xEventGroupCreate();
+    ESP_ERROR_CHECK( esp_event_loop_init(event_handler, NULL) );
+    wifi_init_config_t cfg = WIFI_INIT_CONFIG_DEFAULT();
+    ESP_ERROR_CHECK( esp_wifi_init(&cfg) );
+    wifi_config_t sta_config = {
+        .sta = {
+            .ssid = WIFI_SSID,
+            .password = WIFI_PASSWORD,
+            .bssid_set = false
+        }
+    };
+    ESP_ERROR_CHECK( esp_wifi_set_mode(WIFI_MODE_STA) );
+    ESP_ERROR_CHECK( esp_wifi_set_config(WIFI_IF_STA, &sta_config) );
+    ESP_ERROR_CHECK( esp_wifi_start() );
+}
+
+void initTemplateplus(f_ptr_t ar) {
+	ESP_ERROR_CHECK( nvs_flash_init() );
+
+#if CONFIG_ENABLE_WIFI
+	wifi_init();
+#endif
+
+#if CONFIG_ENABLE_OTA
+    TaskHandle_t otat = otaserver_entry();
+#endif
+
+#ifdef CONFIG_ENABLE_ARDUINO
+    arduino_entry(ar);
+#endif
+
+#if CONFIG_ENABLE_MICROPYTHON
+    micropython_entry();
+#endif
+
+#if CONFIG_ENABLE_WIFI
+#if CONFIG_WIFI_AUTOCONNECT
+	xEventGroupWaitBits(wifi_event_group, WIFI_CONNECTED_BIT, false, true, portMAX_DELAY);
+	vTaskResume(otat);
+#endif
+#endif
+}
