--- firmware/partitions_4M_factory1272K_siffs2736K.csv
@@ -1,8 +1,0 @@
-# -------------------------------------------------------
-# Name,         Type, SubType, Offset,  Size,       Flags
-# -------------------------------------------------------
-nvs,            data, nvs,     0x9000,  16K,
-otadata,	data, ota,     0xd000,	8K,
-phy_init,       data, phy,     0xf000,  4K,
-factory,    	app,  factory, 0x10000, 1272K,
-internalfs,	data, spiffs,  ,        2736K,
