--- firmware/partitions_4M_factory2ota1272K_siffs192K.csv
@@ -1,10 +1,0 @@
-# -------------------------------------------------------
-# Name,         Type, SubType, Offset,  Size,       Flags
-# -------------------------------------------------------
-nvs,            data, nvs,     0x9000,  16K,
-otadata,	data, ota,     0xd000,	8K,
-phy_init,       data, phy,     0xf000,  4K,
-factory,    	app,  factory, 0x10000, 1272K,
-boot1,    	app,  ota_0,   ,        1272K,
-boot2,		app,  ota_1,   ,        1272K,
-internalfs,	data, spiffs,  ,        192K,
