--- main/ino/ino.c
@@ -1,5 +1,0 @@
-static void setup() {
-}
-
-static void loop() {
-}
