--- main/ino/ino.h
@@ -1,8 +1,0 @@
-#ifndef MAIN_INO_INO_H_
-#define MAIN_INO_INO_H_
-
-static void setup();
-
-static void loop();
-
-#endif /* MAIN_INO_INO_H_ */
